Make navbar logout button work
