+++ eztip/src/components/container/NavigationContainer/index.js
@@ -1,0 +1,1 @@
+export { default as NavigationContainer } from "./NavigationContainer";
