--- eztip/src/components/presentational/Navigation/index.js
@@ -1,1 +1,0 @@
-export { default as NavigationContainer } from "./NavigationContainer";
